Accepts the page-name as input from user
The page name would be taken as input from user and top 5 results of the facebook pages based on that name would be shown. 
The user then has the option to choose the page no. among those 5 options and once he enters the no., the web-browser opens up displaying the table with recent 5 posts and top 5 comments on them.